feat: preserve scroll position after card shuffle
Shuffling pokemons after clicking a card causes a layout change
resulting in the scroll position resetting to the top of the screen.
UseLayoutEffect and useRef was added to restore scroll position,
improving UX by helping users keep track of the card they're on.

Fixed scoreboard to corner so that users can keep track of score without
having to scroll all the way to the top.